+++ internal/indexes/scan.go
@@ -1,12 +1,12 @@
 package indexes
 
 type tidScanner interface {
-	Scan() (int, error)
+	Scan() (int64, error)
 }
 
-type tidScannerFunc func() (int, error)
+type tidScannerFunc func() (int64, error)
 
-func (f tidScannerFunc) Scan() (int, error) {
+func (f tidScannerFunc) Scan() (int64, error) {
 	return f()
 }
